requirements.txt
asgiref==3.7.2
celery==5.3.1
Django==5.0.1
sqlparse==0.4.4
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
httpx==0.26.0